+++ frontend/.prettierrc
@@ -1,10 +1,10 @@
 {
-  "tabWidth": 4,
-  "semi": false,
-  "printWidth": 80,
-  "singleQuote": false,
-  "bracketSpacing": true,
-  "trailingComma": "es5",
-  "arrowParens": "always",
-  "plugins": ["prettier-plugin-tailwindcss"]
+    "tabWidth": 4,
+    "semi": true,
+    "printWidth": 80,
+    "singleQuote": false,
+    "bracketSpacing": true,
+    "trailingComma": "es5",
+    "arrowParens": "always",
+    "jsxSingleQuote": false
 }
