--- frontend/app/+not-found.tsx
@@ -1,33 +1,0 @@
-import React from "react";
-import { Link, Stack } from "expo-router";
-import { StyleSheet } from "react-native";
-
-import { ThemedText } from "@/components/ThemedText";
-import { ThemedView } from "@/components/ThemedView";
-
-export default function NotFoundScreen() {
-    return (
-        <React.Fragment>
-            <Stack.Screen options={{ title: "Oops!" }} />
-            <ThemedView style={styles.container}>
-                <ThemedText type="title">This screen doesn't exist.</ThemedText>
-                <Link href="/" style={styles.link}>
-                    <ThemedText type="link">Go to home screen!</ThemedText>
-                </Link>
-            </ThemedView>
-        </React.Fragment>
-    );
-}
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        alignItems: "center",
-        justifyContent: "center",
-        padding: 20,
-    },
-    link: {
-        marginTop: 15,
-        paddingVertical: 15,
-    },
-});
