+++ frontend/app/index.tsx
@@ -1,0 +1,15 @@
+import { Text, View } from "react-native";
+
+export default function Index() {
+    return (
+        <View
+            style={{
+                flex: 1,
+                justifyContent: "center",
+                alignItems: "center",
+            }}
+        >
+            <Text>Edit app/index.tsx to edit this screen.</Text>
+        </View>
+    );
+}
