--- frontend/babel.config.js
@@ -1,9 +1,0 @@
-module.exports = function (api) {
-    api.cache(true);
-    return {
-        presets: [
-            ["babel-preset-expo", { jsxImportSource: "nativewind" }],
-            "nativewind/babel",
-        ],
-    };
-};
