--- frontend/components/HapticTab.tsx
@@ -1,18 +1,0 @@
-import * as Haptics from "expo-haptics";
-import { PlatformPressable } from "@react-navigation/elements";
-import { BottomTabBarButtonProps } from "@react-navigation/bottom-tabs";
-
-export function HapticTab(props: BottomTabBarButtonProps) {
-    return (
-        <PlatformPressable
-            {...props}
-            onPressIn={(ev) => {
-                if (process.env.EXPO_OS === "ios") {
-                    // Add a soft haptic feedback when pressing down on the tabs.
-                    Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
-                }
-                props.onPressIn?.(ev);
-            }}
-        />
-    );
-}
