--- frontend/components/ThemedView.tsx
@@ -1,22 +1,0 @@
-import { View, type ViewProps } from "react-native";
-
-import { useThemeColor } from "@/hooks/useThemeColor";
-
-export type ThemedViewProps = ViewProps & {
-    lightColor?: string;
-    darkColor?: string;
-};
-
-export function ThemedView({
-    style,
-    lightColor,
-    darkColor,
-    ...otherProps
-}: ThemedViewProps) {
-    const backgroundColor = useThemeColor(
-        { light: lightColor, dark: darkColor },
-        "background"
-    );
-
-    return <View style={[{ backgroundColor }, style]} {...otherProps} />;
-}
