--- frontend/components/__tests__/ThemedText-test.tsx
@@ -1,12 +1,0 @@
-import * as React from "react";
-import renderer from "react-test-renderer";
-
-import { ThemedText } from "../ThemedText";
-
-it(`renders correctly`, () => {
-    const tree = renderer
-        .create(<ThemedText>Snapshot test!</ThemedText>)
-        .toJSON();
-
-    expect(tree).toMatchSnapshot();
-});
