--- frontend/global.css
@@ -1,3 +1,0 @@
-@tailwind base;
-@tailwind utilities;
-@tailwind components;
