--- frontend/metro.config.js
@@ -1,6 +1,0 @@
-const { withNativeWind } = require("nativewind/metro");
-const { getDefaultConfig } = require("expo/metro-config");
-
-const config = getDefaultConfig(__dirname);
-
-module.exports = withNativeWind(config, { input: "./global.css" });
