--- frontend/nativewind-env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="nativewind/types" />
